Refactored the app file structure and added a cache for storing data
